Ensure received emails with no chain are signed correctly
If OpenARC is used to verify an inbound email (with no existing chain),
then if the email is modified and re-signed, then the chain is broken.
This is because OpenARC carries forward the existing none verification
status, rather than changing it to a pass as would be expected from this
scenario functioning correctly.

This commit ensures that a message that was received with none which now
has one chain has its cv changed to pass instead of remaining at none.